Fix segfault in linux
When a large operation happens in linux sometimes a value in mWDtoDirNode
could get erased without the pointer that lives in the map being freed.

This led to some strange behavior on linux that caused a giant string to
try to be created since the pointer was pointing at garbage memory now.

The fix means that we now free the pointer when needed and do a check when
we're processing directory events for watch descriptors that we don't track
anymore.

As a consequence sometimes we will lose events but the application won't
crash anymore and we won't have memory leaks. Also less crying.