docs: Indent options' children properly
The children of `options` are indented with only one space.  In the rendered Markdown (at least on Github), they appear at the same level as `options` itself, which makes it unclear that they are members of `options`.  Indent them with two spaces so they also appear indented in the rendered Markdown.