Address undefined behavior
TIL when running the optimizer while compiling, if you don't include a mutex in a function with a variable that is shared across threads, it basically tells the optimizer that it's safe to assume that this method is single threaded.

Running: the code provided here in g++ with -O3 enabled will cause this program to loop infinitely. 
https://devblogs.microsoft.com/oldnewthing/20180924-00/?p=99805