linux: support simple ignore patterns
This is a first pass at implementing some mechanism to filter paths.
Linux only.

Some pattern examples:

- `/absolute/path[/][*]`
- `./path/parts[/][*]`
- `**/path/parts[/][*]`
- `*path/parts[/][*]`
- `path/parts[/][*]`

The general idea is to do simple string matching and convert patterns
into path parts that can then be matched against a real path.

For example, `**/foo` gets converted to `/foo/` and we will try to find
this subtring in the following paths:

- `/a/b/foo/c/` matches.
- `/a/b/foobar/c/` does not match.

In order for the pattern to match `/a/b/foobar/c/` we need to write
`**/foo*` which will be converted to `/foo`. We now have:

- `/a/b/foo/c` matches.
- `/a/b/foobar/c` matches.

This is because we are not required to find `/` right after `/foo`.
Depending on if we use a `/` character or not, we can match path parts
strictly or not.

The rules for converting patterns are as follow:

- If a pattern starts with `**/` we replace this prefix by `/` and will
match the string anywhere.
- If a pattern starts with `*` we do not prefix it with `/` and will
match the string anywhere.
- If a pattern starts with `./` we replace this prefix by the directory
being watched by the current watcher and will match the string from
root.
- If a pattern starts with `/` we will match the string from root.
- If a pattern starts with none of the above, we prefix with `/` and
match the string anywhere.

Then we apply trailing conversion logic:

- If a pattern contains a `*` we remove this character plus everything
that follows.
- If a pattern ends with `/` we ensure there's only one trailing `/`.
- If a pattern does not finish by either `*` or `/` we append `/`.

The final filters are triaged between two vectors: `mRootFilters` and
`mFilters` in order to match starting from the root or to match
anywhere, respectively.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>