attach inotify watcher before scandir
There is a window where some created folders might get ignored
completely. It happens between the time where the tree is created and
the time where the tree is iterated over to allocate watchers.

This commit hoist the inotify watching before doing scandir. This way
there is no window where a directory might be created after the initial
scandir and missed, as the inotify watcher should pick it up.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>