gc: fix instance allocation counting
A class destructor is not called if an exception is thrown when
executing the constructor. This caused an issue where if an error
occured during construction of NSFW objects, the instance coutn would
never be decremented.

This commit moves the increment step at the end of the constructor,
after which nothing can error out.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>